+++ src/commands/onboard-non-interactive.gateway.test.ts
@@ -1,20 +1,37 @@
-// Non-interactive gateway onboarding tests cover local/remote setup, auth, daemon install, and config writes.
+// Non-interactive gateway onboarding tests cover local/remote setup, daemon install, and config writes.
+// Gateway auth-token storage has its own suite in onboard-non-interactive.gateway-auth-token.test.ts.
 import fs from "node:fs/promises";
 import path from "node:path";
 import { afterAll, afterEach, beforeAll, describe, expect, it, vi } from "vitest";
-import type { ConfigFileSnapshot, OpenClawConfig } from "../config/types.openclaw.js";
+import type { OpenClawConfig } from "../config/types.openclaw.js";
 import type { RuntimeEnv } from "../runtime.js";
 import { makeTempWorkspace } from "../test-helpers/workspace.js";
 import { setTestEnvValue } from "../test-utils/env.js";
+import {
+  capturedReplaceConfigFileCalls,
+  configWritePluginLeaseDepths,
+  ensureWorkspaceAndSessionsMock,
+  gatewayReachableState,
+  gatewayServiceMock,
+  getPseudoPort,
+  healthCommandMock,
+  installGatewayDaemonNonInteractiveMock,
+  loadGatewayOnboardModules,
+  readConfigFileSnapshotMock,
+  readLastGatewayErrorLineMock,
+  readTestConfig,
+  resolveInstallDaemonGatewayHealthTiming,
+  resolveTestConfigPath,
+  runNonInteractiveSetup,
+  runtime,
+  testConfigStore,
+} from "./onboard-non-interactive.gateway.test-mocks.js";
 import {
   createOnboardGatewayTimeoutCapture,
   createOnboardJsonCaptureRuntime,
   createOnboardLocalDaemonOptions,
   createOnboardStateDirHarness,
-  createOnboardTestConfigStore,
-  createThrowingRuntime,
   expectOnboardLocalJsonSetupFailure,
-  mockOnboardingAgent,
   prepareOnboardGatewayTestEnv,
   readOnboardFirstMockCall,
   runOnboardLocalDaemonSetup,
@@ -23,202 +40,7 @@ import type {
   OnboardEnsureWorkspaceOptions,
   OnboardGatewayHealthCall,
   OnboardHealthCommandCall,
-  WaitForGatewayReachableMock,
 } from "./onboard-non-interactive.test-helpers.js";
-import type { installGatewayDaemonNonInteractive } from "./onboard-non-interactive/local/daemon-install.js";
-
-const ensureWorkspaceAndSessionsMock = vi.fn(async (..._args: unknown[]) => {});
-const {
-  configStore: testConfigStore,
-  resolveConfigPath: resolveTestConfigPath,
-  readConfig: readTestConfig,
-  readSnapshot: readTestConfigSnapshot,
-} = createOnboardTestConfigStore();
-const readConfigFileSnapshotMock = vi.hoisted(() => vi.fn<() => Promise<ConfigFileSnapshot>>());
-const pluginLifecycleLeaseState = vi.hoisted(() => ({ depth: 0 }));
-const configWritePluginLeaseDepths: number[] = [];
-type InstallGatewayDaemonResult = Awaited<ReturnType<typeof installGatewayDaemonNonInteractive>>;
-const installGatewayDaemonNonInteractiveMock = vi.hoisted(() =>
-  vi.fn(async (): Promise<InstallGatewayDaemonResult> => ({ installed: true })),
-);
-const healthCommandMock = vi.hoisted(() => vi.fn(async () => {}));
-const gatewayServiceMock = vi.hoisted(() => ({
-  label: "LaunchAgent",
-  loadedText: "loaded",
-  isLoaded: vi.fn(async () => true),
-  readRuntime: vi.fn(async () => ({
-    status: "running",
-    state: "active",
-    pid: 4242,
-  })),
-}));
-const readLastGatewayErrorLineMock = vi.hoisted(() =>
-  vi.fn(async () => "Gateway failed to start: required secrets are unavailable."),
-);
-let waitForGatewayReachableMock: WaitForGatewayReachableMock;
-
-readConfigFileSnapshotMock.mockImplementation(async () => readTestConfigSnapshot());
-
-vi.mock("../config/io.js", () => ({
-  createConfigIO: () => ({
-    configPath: resolveTestConfigPath(),
-  }),
-  loadConfig: () => readTestConfig(),
-  readConfigFileSnapshot: readConfigFileSnapshotMock,
-}));
-
-vi.mock("../plugins/plugin-lifecycle-lease.js", () => ({
-  withPluginLifecycleLease: async (
-    _options: unknown,
-    run: (lease: {
-      databasePath: string;
-      signal: AbortSignal;
-      assertOwned: () => void;
-      assertOwnedInTransaction: () => void;
-    }) => Promise<unknown>,
-  ) => {
-    pluginLifecycleLeaseState.depth += 1;
-    try {
-      return await run({
-        databasePath: path.join(path.dirname(resolveTestConfigPath()), "openclaw.sqlite"),
-        signal: new AbortController().signal,
-        assertOwned: () => {},
-        assertOwnedInTransaction: () => {},
-      });
-    } finally {
-      pluginLifecycleLeaseState.depth -= 1;
-    }
-  },
-}));
-
-const capturedReplaceConfigFileCalls: Array<{
-  nextConfig: OpenClawConfig;
-  writeOptions?: { allowConfigSizeDrop?: boolean; unsetPaths?: string[][] };
-}> = [];
-
-vi.mock("../config/config.js", async (importActual) => {
-  const actual = await importActual<typeof import("../config/config.js")>();
-  return {
-    replaceConfigFile: async ({
-      nextConfig,
-      writeOptions,
-    }: {
-      nextConfig: OpenClawConfig;
-      writeOptions?: { allowConfigSizeDrop?: boolean; unsetPaths?: string[][] };
-    }) => {
-      configWritePluginLeaseDepths.push(pluginLifecycleLeaseState.depth);
-      capturedReplaceConfigFileCalls.push({
-        nextConfig,
-        ...(writeOptions ? { writeOptions } : {}),
-      });
-      testConfigStore.set(resolveTestConfigPath(), nextConfig);
-    },
-    resolveConfigWriteAfterWrite: actual.resolveConfigWriteAfterWrite,
-    resolveGatewayPort: (cfg: OpenClawConfig) => cfg.gateway?.port ?? 18789,
-    transformConfigFileWithRetry: async (
-      params: Parameters<typeof import("../config/config.js").transformConfigFileWithRetry>[0],
-    ) => {
-      const snapshot = await readConfigFileSnapshotMock();
-      const previousHash = snapshot.hash ?? null;
-      const transformed = await params.transform(snapshot.sourceConfig, {
-        snapshot,
-        previousHash,
-        attempt: 0,
-      });
-      const committed = await params.commit!({
-        nextConfig: transformed.nextConfig,
-        snapshot,
-        ...(previousHash ? { baseHash: previousHash } : {}),
-        writeOptions: params.writeOptions,
-        afterWrite: { mode: "auto" },
-      });
-      return { nextConfig: committed.config };
-    },
-  };
-});
-
-vi.mock("./onboard-agent.js", () => ({ ensureOnboardingAgent: mockOnboardingAgent }));
-
-vi.mock("./onboard-helpers.js", () => {
-  const normalizeGatewayTokenInput = (value: unknown): string => {
-    if (typeof value !== "string") {
-      return "";
-    }
-    const trimmed = value.trim();
-    return trimmed === "undefined" || trimmed === "null" ? "" : trimmed;
-  };
-  return {
-    DEFAULT_WORKSPACE: "/tmp/openclaw-workspace",
-    applyWizardMetadata: (cfg: unknown) => cfg,
-    ensureWorkspaceAndSessions: ensureWorkspaceAndSessionsMock,
-    normalizeGatewayTokenInput,
-    randomToken: () => "tok_generated_gateway_test_token",
-    resolveControlUiLinks: ({ port }: { port: number }) => ({
-      httpUrl: `http://127.0.0.1:${port}`,
-      wsUrl: `ws://127.0.0.1:${port}`,
-    }),
-    resolveLocalControlUiProbeLinks: ({ port }: { port: number }) => ({
-      httpUrl: `http://127.0.0.1:${port}`,
-      wsUrl: `ws://127.0.0.1:${port}`,
-    }),
-    waitForGatewayReachable: (params: {
-      url: string;
-      token?: string;
-      password?: string;
-      deadlineMs?: number;
-      probeTimeoutMs?: number;
-    }) => waitForGatewayReachableMock?.(params) ?? Promise.resolve({ ok: true }),
-  };
-});
-
-vi.mock("./onboard-non-interactive/local/daemon-install.js", () => ({
-  installGatewayDaemonNonInteractive: installGatewayDaemonNonInteractiveMock,
-}));
-
-vi.mock("./health.js", () => ({
-  healthCommandNonExiting: healthCommandMock,
-}));
-
-vi.mock("../daemon/service.js", () => ({
-  readGatewayServiceState: async () => {
-    const [loadState, runtime] = await Promise.all([
-      gatewayServiceMock
-        .isLoaded()
-        .then((loaded) =>
-          loaded ? ({ status: "loaded" } as const) : ({ status: "not-loaded" } as const),
-        )
-        .catch((error: unknown) => ({ status: "unknown" as const, detail: String(error) })),
-      gatewayServiceMock.readRuntime(),
-    ]);
-    return {
-      installed: true,
-      loadState,
-      running: runtime.status === "running",
-      env: {},
-      command: null,
-      runtime,
-    };
-  },
-  resolveGatewayService: () => gatewayServiceMock,
-}));
-
-vi.mock("../daemon/diagnostics.js", () => ({
-  readLastGatewayErrorLine: readLastGatewayErrorLineMock,
-}));
-
-let runNonInteractiveSetup: typeof import("./onboard-non-interactive.js").runNonInteractiveSetup;
-let resolveInstallDaemonGatewayHealthTiming: typeof import("./onboard-non-interactive/local.test-support.js").resolveInstallDaemonGatewayHealthTiming;
-
-async function loadGatewayOnboardModules(): Promise<void> {
-  vi.resetModules();
-  ({ runNonInteractiveSetup } = await import("./onboard-non-interactive.js"));
-  ({ resolveInstallDaemonGatewayHealthTiming } =
-    await import("./onboard-non-interactive/local.test-support.js"));
-}
-
-const getPseudoPort = (base: number): number => base + (process.pid % 1000);
-
-const runtime = createThrowingRuntime();
 
 describe("onboard (non-interactive): gateway and remote auth", () => {
   let envSnapshot: ReturnType<typeof prepareOnboardGatewayTestEnv>;
@@ -241,7 +63,7 @@ describe("onboard (non-interactive): gateway and remote auth", () => {
   });
 
   afterEach(() => {
-    waitForGatewayReachableMock = undefined;
+    gatewayReachableState.mock = undefined;
     testConfigStore.clear();
     capturedReplaceConfigFileCalls.length = 0;
     configWritePluginLeaseDepths.length = 0;
@@ -425,58 +247,6 @@ describe("onboard (non-interactive): gateway and remote auth", () => {
       expect(onboardWrite?.nextConfig.plugins?.installs).toBeUndefined();
       expect(onboardWrite?.writeOptions?.unsetPaths).toEqual([["plugins", "installs"]]);
       expect(onboardWrite?.writeOptions?.allowConfigSizeDrop).toBe(false);
-    });
-  }, 60_000);
-
-  it("writes gateway token auth into config", async () => {
-    await withStateDir("state-noninteractive-", async (stateDir) => {
-      const token = "tok_test_123";
-      const workspace = path.join(stateDir, "openclaw");
-      testConfigStore.set(resolveTestConfigPath(), {
-        gateway: {
-          bind: "lan",
-          auth: { mode: "password", password: "test-password" },
-          tailscale: { mode: "serve" },
-        },
-      } as OpenClawConfig);
-
-      await runNonInteractiveSetup(
-        {
-          nonInteractive: true,
-          mode: "local",
-          workspace,
-          authChoice: "skip",
-          skipSkills: true,
-          skipHealth: true,
-          installDaemon: false,
-          gatewayBind: "loopback",
-          gatewayAuth: "token",
-          gatewayToken: token,
-          tailscale: "off",
-        },
-        runtime,
-      );
-
-      const cfg = readTestConfig() as {
-        gateway?: {
-          mode?: string;
-          bind?: string;
-          auth?: { mode?: string; token?: string };
-          tailscale?: { mode?: string };
-        };
-        agents?: { defaults?: { workspace?: string } };
-        tools?: { profile?: string };
-        hooks?: { internal?: { entries?: Record<string, { enabled?: boolean }> } };
-      };
-
-      expect(cfg?.agents?.defaults?.workspace).toBe(workspace);
-      expect(cfg?.gateway?.mode).toBe("local");
-      expect(cfg?.gateway?.bind).toBe("loopback");
-      expect(cfg?.tools?.profile).toBe("coding");
-      expect(cfg?.gateway?.auth?.mode).toBe("token");
-      expect(cfg?.gateway?.auth?.token).toBe(token);
-      expect(cfg?.gateway?.tailscale).toEqual({ mode: "off" });
-      expect(cfg?.hooks?.internal?.entries?.["session-memory"]).toEqual({ enabled: true });
     });
   }, 60_000);
 
@@ -687,7 +457,7 @@ describe("onboard (non-interactive): gateway and remote auth", () => {
 
   it("completes explicit no-daemon setup when no gateway is listening", async () => {
     await withStateDir("state-local-health-hint-", async (stateDir) => {
-      waitForGatewayReachableMock = vi.fn(async () => ({
+      gatewayReachableState.mock = vi.fn(async () => ({
         ok: false,
         detail: "connect ECONNREFUSED 127.0.0.1:18789",
       }));
@@ -706,7 +476,7 @@ describe("onboard (non-interactive): gateway and remote auth", () => {
 
   it("still fails when an existing gateway is expected but unreachable", async () => {
     await withStateDir("state-local-health-required-", async (stateDir) => {
-      waitForGatewayReachableMock = vi.fn(async () => ({
+      gatewayReachableState.mock = vi.fn(async () => ({
         ok: false,
         detail: "connect ECONNREFUSED 127.0.0.1:18789",
       }));
@@ -725,7 +495,7 @@ describe("onboard (non-interactive): gateway and remote auth", () => {
   it("uses a longer health deadline when daemon install was requested", async () => {
     await withStateDir("state-local-daemon-health-", async (stateDir) => {
       const captured = createOnboardGatewayTimeoutCapture();
-      waitForGatewayReachableMock = captured.mock;
+      gatewayReachableState.mock = captured.mock;
 
       await runOnboardLocalDaemonSetup({ runSetup: runNonInteractiveSetup, stateDir, runtime });
 
@@ -744,7 +514,7 @@ describe("onboard (non-interactive): gateway and remote auth", () => {
   it("passes pinned gateway auth through non-interactive health checks", async () => {
     await withStateDir("state-local-daemon-health-auth-", async (stateDir) => {
       const token = "tok_noninteractive_health";
-      waitForGatewayReachableMock = vi.fn(async () => ({ ok: true }));
+      gatewayReachableState.mock = vi.fn(async () => ({ ok: true }));
 
       await runNonInteractiveSetup(
         {
@@ -756,7 +526,7 @@ describe("onboard (non-interactive): gateway and remote auth", () => {
       );
 
       const [gatewayHealthCall] = readOnboardFirstMockCall(
-        waitForGatewayReachableMock,
+        gatewayReachableState.mock,
         "waitForGatewayReachable",
       ) as [OnboardGatewayHealthCall];
       expect(gatewayHealthCall.token).toBe(token);
@@ -839,7 +609,7 @@ describe("onboard (non-interactive): gateway and remote auth", () => {
 
   it("emits structured JSON diagnostics when daemon health fails", async () => {
     await withStateDir("state-local-daemon-health-json-fail-", async (stateDir) => {
-      waitForGatewayReachableMock = vi.fn(async () => ({
+      gatewayReachableState.mock = vi.fn(async () => ({
         ok: false,
         detail: "gateway closed (1006 abnormal closure (no close frame)): no close reason",
       }));
@@ -886,7 +656,7 @@ describe("onboard (non-interactive): gateway and remote auth", () => {
 
   it("emits structured JSON failure when a reachable gateway fails its health check", async () => {
     await withStateDir("state-local-daemon-health-exit-json-", async (stateDir) => {
-      waitForGatewayReachableMock = vi.fn(async () => ({ ok: true }));
+      gatewayReachableState.mock = vi.fn(async () => ({ ok: true }));
       healthCommandMock.mockImplementationOnce(async (...args: unknown[]) => {
         // healthCommand prints its reachable-gateway diagnostic before its
         // CLI-style exit; the capture runtime must keep it off JSON stdout.
@@ -924,7 +694,7 @@ describe("onboard (non-interactive): gateway and remote auth", () => {
 
   it("routes thrown health-check errors through the onboarding failure owner", async () => {
     await withStateDir("state-local-health-failure-text-", async (stateDir) => {
-      waitForGatewayReachableMock = vi.fn(async () => ({ ok: true }));
+      gatewayReachableState.mock = vi.fn(async () => ({ ok: true }));
       healthCommandMock.mockRejectedValueOnce(new Error("health request timed out"));
 
       await expect(
@@ -935,7 +705,7 @@ describe("onboard (non-interactive): gateway and remote auth", () => {
 
   it("preserves unknown service inspection in JSON diagnostics", async () => {
     await withStateDir("state-local-daemon-health-unknown-", async (stateDir) => {
-      waitForGatewayReachableMock = vi.fn(async () => ({
+      gatewayReachableState.mock = vi.fn(async () => ({
         ok: false,
         detail: "connect ECONNREFUSED 127.0.0.1:18789",
       }));
@@ -963,7 +733,7 @@ describe("onboard (non-interactive): gateway and remote auth", () => {
 
   it("classifies daemon health ECONNREFUSED failures with a recovery command", async () => {
     await withStateDir("state-local-daemon-health-refused-", async (stateDir) => {
-      waitForGatewayReachableMock = vi.fn(async () => ({
+      gatewayReachableState.mock = vi.fn(async () => ({
         ok: false,
         detail: "connect ECONNREFUSED 127.0.0.1:18789",
       }));
@@ -991,135 +761,6 @@ describe("onboard (non-interactive): gateway and remote auth", () => {
       expect(parsed.phase).toBe("gateway-health");
       expect(parsed.classification).toBe("service-stopped");
       expect(parsed.hints).toContain("Fix: run `openclaw gateway restart`.");
-    });
-  }, 60_000);
-
-  it("auto-generates token auth when binding LAN and persists the token", async () => {
-    if (process.platform === "win32") {
-      // Windows runner occasionally drops the temp config write in this flow; skip to keep CI green.
-      return;
-    }
-    await withStateDir("state-lan-", async (stateDir) => {
-      setTestEnvValue("OPENCLAW_STATE_DIR", stateDir);
-      setTestEnvValue("OPENCLAW_CONFIG_PATH", path.join(stateDir, "openclaw.json"));
-
-      const port = getPseudoPort(40_000);
-      const workspace = path.join(stateDir, "openclaw");
-
-      await runNonInteractiveSetup(
-        {
-          nonInteractive: true,
-          mode: "local",
-          workspace,
-          authChoice: "skip",
-          skipSkills: true,
-          skipHealth: true,
-          installDaemon: false,
-          gatewayPort: port,
-          gatewayBind: "lan",
-        },
-        runtime,
-      );
-
-      const cfg = readTestConfig() as {
-        gateway?: {
-          bind?: string;
-          port?: number;
-          auth?: { mode?: string; token?: string };
-        };
-      };
-
-      expect(cfg.gateway?.bind).toBe("lan");
-      expect(cfg.gateway?.port).toBe(port);
-      expect(cfg.gateway?.auth?.mode).toBe("token");
-      expect((cfg.gateway?.auth?.token ?? "").length).toBeGreaterThan(8);
-    });
-  }, 60_000);
-
-  it("keeps the generated gateway token out of config under --secret-input-mode ref", async () => {
-    if (process.platform === "win32") {
-      // Matches the LAN case above: the Windows runner drops this flow's temp config write.
-      return;
-    }
-    await withStateDir("state-token-ref-", async (stateDir) => {
-      setTestEnvValue("OPENCLAW_STATE_DIR", stateDir);
-      setTestEnvValue("OPENCLAW_CONFIG_PATH", path.join(stateDir, "openclaw.json"));
-
-      const port = getPseudoPort(41_000);
-
-      await runNonInteractiveSetup(
-        {
-          nonInteractive: true,
-          mode: "local",
-          workspace: path.join(stateDir, "openclaw"),
-          authChoice: "skip",
-          skipSkills: true,
-          skipHealth: true,
-          installDaemon: false,
-          gatewayPort: port,
-          secretInputMode: "ref",
-        },
-        runtime,
-      );
-
-      const cfg = readTestConfig() as {
-        gateway?: { auth?: { mode?: string; token?: unknown } };
-      };
-      expect(cfg.gateway?.auth?.mode).toBe("token");
-      expect(cfg.gateway?.auth?.token).toEqual({
-        source: "store",
-        provider: "default",
-        id: "OPENCLAW_GATEWAY_TOKEN",
-      });
-
-      // A ref persisted without its value would leave the gateway unauthenticatable.
-      const { readSecretStoreValue } = await import("../secrets/store/secret-store.js");
-      const stored = readSecretStoreValue({
-        scope: { kind: "team" },
-        name: "OPENCLAW_GATEWAY_TOKEN",
-      });
-      expect(stored.ok).toBe(true);
-      expect(stored.ok && stored.value.length).toBeGreaterThan(8);
-    });
-  }, 60_000);
-
-  it("references an ambient gateway token by env instead of copying it into the store", async () => {
-    if (process.platform === "win32") {
-      // Matches the LAN case above: the Windows runner drops this flow's temp config write.
-      return;
-    }
-    await withStateDir("state-token-ref-env-", async (stateDir) => {
-      setTestEnvValue("OPENCLAW_STATE_DIR", stateDir);
-      setTestEnvValue("OPENCLAW_CONFIG_PATH", path.join(stateDir, "openclaw.json"));
-      setTestEnvValue("OPENCLAW_GATEWAY_TOKEN", "ambient-gateway-token");
-
-      await runNonInteractiveSetup(
-        {
-          nonInteractive: true,
-          mode: "local",
-          workspace: path.join(stateDir, "openclaw"),
-          authChoice: "skip",
-          skipSkills: true,
-          skipHealth: true,
-          installDaemon: false,
-          gatewayPort: getPseudoPort(42_000),
-          secretInputMode: "ref",
-        },
-        runtime,
-      );
-
-      const cfg = readTestConfig() as { gateway?: { auth?: { token?: unknown } } };
-      expect(cfg.gateway?.auth?.token).toEqual({
-        source: "env",
-        provider: "default",
-        id: "OPENCLAW_GATEWAY_TOKEN",
-      });
-
-      // A store copy would silently outlive a later rotation of the env var.
-      const { readSecretStoreValue } = await import("../secrets/store/secret-store.js");
-      expect(
-        readSecretStoreValue({ scope: { kind: "team" }, name: "OPENCLAW_GATEWAY_TOKEN" }).ok,
-      ).toBe(false);
     });
   }, 60_000);
 });
